+++ CHANGELOG.md
@@ -23,6 +23,9 @@
   odd number of quotes when being formatted. The previous implementation would
   use an even number, which is invalid PRQL. (@max-sixty, #4850)
 
+- A few more keywords are quoted, such as `user`, which is a reserved keyword in
+  PostgreSQL. (@max-sixty)
+
 **Documentation**:
 
 **Web**: